refactor: IntegrationTest abstract 클래스 추가
- test용 application.properties 추가
- data.sql 추가
- schema.sql 추가
