+++ src/test/java/kitchenpos/ApplicationTest.java
@@ -5,6 +5,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class ApplicationTest {
+
     @Test
     void contextLoads() {
     }
