+++ src/test/resources/application.properties
@@ -1,0 +1,6 @@
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+spring.h2.console.enabled=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
+spring.flyway.enabled=false
+spring.jpa.hibernate.ddl-auto=none
